feat(example): add scenario for prove-core
