+++ examples/fibonacci/script/src/operator/steps.rs
@@ -1,10 +1,9 @@
+use crate::common;
 use crate::common::types::{
     ChallengerType, CheckpointType, CommitmentType, PublicValueStreamType, PublicValuesType,
     RecordType,
 };
 use crate::ProveArgs;
-use crate::{common, PublicValuesTuple};
-use alloy_sol_types::SolType;
 use anyhow::Result;
 use p3_baby_bear::BabyBear;
 use sp1_core::stark::MachineRecord;
@@ -14,7 +13,7 @@ use sp1_core::{
     utils::{BabyBearPoseidon2, SP1CoreProverError},
 };
 use sp1_prover::{SP1CoreProof, SP1CoreProofData, SP1ProofWithMetadata};
-use sp1_sdk::{SP1Proof, SP1ProofWithPublicValues, SP1PublicValues};
+use sp1_sdk::SP1PublicValues;
 
 fn operator_split_into_checkpoints(
     runtime: &mut Runtime,
@@ -151,30 +150,4 @@ pub fn construct_sp1_core_proof_impl(
     };
 
     Ok(sp1_core_proof)
-}
-
-pub fn construct_core_proof(
-    args: &Vec<u8>,
-    core_proof: &Vec<u8>,
-) -> Result<SP1ProofWithPublicValues> {
-    let args_obj = ProveArgs::from_slice(args.as_slice());
-    let core_proof_obj: SP1CoreProof = bincode::deserialize(core_proof).unwrap();
-
-    let (client, _, _, vk) = common::init_client(args_obj);
-
-    let proof = SP1ProofWithPublicValues {
-        proof: SP1Proof::Core(core_proof_obj.proof.0),
-        stdin: core_proof_obj.stdin,
-        public_values: core_proof_obj.public_values,
-        sp1_version: client.prover.version().to_string(),
-    };
-
-    client.verify(&proof, &vk).expect("failed to verify proof");
-    tracing::info!("Successfully generated core-proof(verified)");
-
-    let (_, _, fib_n) =
-        PublicValuesTuple::abi_decode(proof.public_values.as_slice(), false).unwrap();
-    tracing::info!("Public Input: {}", fib_n);
-
-    Ok(proof)
 }
