+++ examples/fibonacci/script/src/scenario/mod.rs
@@ -1,0 +1,1 @@
+pub mod core_prove;
